requirements.txt
PyMuPDF>=1.24.0
pydantic>=2.0
pyyaml>=6.0
streamlit>=1.50.0
anthropic>=0.30.0
pytest>=8.0
pytest-cov>=5.0
